backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
sqlalchemy==2.0.27
python-dotenv==1.0.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
sentence-transformers==2.5.1
scikit-learn==1.4.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
PyMuPDF==1.23.21
bs4==0.0.1 